+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [0.1.0] - 2020-01-06
-* Initial release
+
+- Initial release
 
 ## [0.2.0] - 2020-01-06
-* Change in Example and Description
+
+- Change in Example and Description
+
+## [0.3.0] - 2020-01-22
+
+- Now you can preload list view with `initialData`